sql: extract zone config utilities to pkg/sql/zoneconfig
Extracted zone configuration functions from pkg/sql to new
pkg/sql/zoneconfig package. Core functions like zoneconfig.Complete(),
zoneconfig.GetInTxn(), and zoneconfig.GetHydratedFor*() are now in the
new package. Planner-dependent code remains in pkg/sql/zone_config_planner.go.

Informs #79357

Release note: None
