+++ pkg/BUILD.bazel
@@ -707,6 +707,7 @@ ALL_TESTS = [
     "//pkg/sql/vecindex/vecpb:vecpb_test",
     "//pkg/sql/vecindex/vecstore:vecstore_test",
     "//pkg/sql/vecindex:vecindex_test",
+    "//pkg/sql/zoneconfig:zoneconfig_test",
     "//pkg/sql:sql_disallowed_imports_test",
     "//pkg/sql:sql_test",
     "//pkg/storage/disk:disk_test",
@@ -2602,6 +2603,8 @@ GO_TARGETS = [
     "//pkg/sql/vecindex:vecindex",
     "//pkg/sql/vecindex:vecindex_test",
     "//pkg/sql/vtable:vtable",
+    "//pkg/sql/zoneconfig:zoneconfig",
+    "//pkg/sql/zoneconfig:zoneconfig_test",
     "//pkg/sql:sql",
     "//pkg/sql:sql_test",
     "//pkg/storage/disk:disk",
