+++ pkg/spanconfig/spanconfigsqltranslator/BUILD.bazel
@@ -14,10 +14,10 @@ go_library(
         "//pkg/roachpb",
         "//pkg/settings/cluster",
         "//pkg/spanconfig",
-        "//pkg/sql",
         "//pkg/sql/catalog",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/descs",
+        "//pkg/sql/zoneconfig",
         "@com_github_cockroachdb_errors//:errors",
     ],
 )
