+++ pkg/sql/zoneconfig/zone_config_helper.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package sql
+package zoneconfig
 
 import (
 	"context"
